Fix for global.lax exposing the entire class
