--- lib/lax.tsbuildinfo
@@ -1,28 +1,0 @@
-{
-  "bundle": {
-    "commonSourceDirectory": "../src",
-    "sourceFiles": [
-      "../src/types.ts",
-      "../src/lax.ts"
-    ],
-    "js": {
-      "sections": [
-        {
-          "pos": 0,
-          "end": 33586,
-          "kind": "text"
-        }
-      ]
-    },
-    "dts": {
-      "sections": [
-        {
-          "pos": 0,
-          "end": 6021,
-          "kind": "text"
-        }
-      ]
-    }
-  },
-  "version": "4.1.3"
-}
